// migrations/snapshots/cb77ab94d987a3fc92b6b872ac2ebe03b83365b4a217074127fd7694c3b3f921/contract.d.ts
export type StorageHash =
  StorageHashBase<'cb77ab94d987a3fc92b6b872ac2ebe03b83365b4a217074127fd7694c3b3f921'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly user: {
              columns: {
                readonly id: {
                  readonly nativeType: 'int4';
                  readonly codecId: 'pg/int4@1';
                  readonly nullable: false;
                  readonly default: {
                    readonly kind: 'function';
                    readonly expression: 'autoincrement()';
                  };
                };
                readonly email: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
                readonly username: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
                readonly createdAt: {
                  readonly nativeType: 'timestamptz';
                  readonly codecId: 'pg/timestamptz-string@1';
                  readonly nullable: false;
                  readonly default: { readonly kind: 'function'; readonly expression: 'now()' };
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [{ readonly columns: readonly ['email'] }];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly user: { readonly namespace: 'public' & NamespaceId; readonly model: 'User' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly User: {
            readonly fields: {
              readonly id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/int4@1' };
              };
              readonly email: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly username: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly createdAt: {
                readonly nullable: false;
                readonly type: {
                  readonly kind: 'scalar';
                  readonly codecId: 'pg/timestamptz-string@1';
                };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'user';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly id: { readonly column: 'id' };
                readonly email: { readonly column: 'email' };
                readonly username: { readonly column: 'username' };
                readonly createdAt: { readonly column: 'createdAt' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly checkConstraint: true;
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;